+++ packages/permissionless/accounts/kernel/toKernelSmartAccount.ts
@@ -224,9 +224,13 @@ const getDefaultAddresses = ({
 }
 
 export const getEcdsaRootIdentifierForKernelV3 = (
-    validatorAddress: Address
+    validatorAddress: Address,
+    eip7702 = false
 ) => {
-    return concatHex([VALIDATOR_TYPE.VALIDATOR, validatorAddress])
+    return concatHex([
+        eip7702 ? VALIDATOR_TYPE.EIP7702 : VALIDATOR_TYPE.VALIDATOR,
+        eip7702 ? "0x" : validatorAddress
+    ])
 }
 
 /**
@@ -416,7 +420,7 @@ export type ToKernelSmartAccountParameters<
           index?: never
           factoryAddress?: never
           metaFactoryAddress?: never
-          accountLogicAddress?: never
+          accountLogicAddress?: Address
           validatorAddress?: never
           nonceKey?: never
           useMetaFactory?: never
@@ -724,8 +728,14 @@ export async function toKernelSmartAccount<
             return this.signMessage({ message: hash })
         },
         async signMessage({ message }) {
-            if (eip7702) {
-                throw new Error("Kernel with EIP-7702 isn't 1271 compliant")
+            if (
+                "isDeployed" in this &&
+                !(await (this as any).isDeployed()) &&
+                eip7702
+            ) {
+                throw new Error(
+                    "Kernel with EIP-7702 isn't 1271 compliant before delegation."
+                )
             }
 
             const signature = await signMessage({
@@ -733,7 +743,8 @@ export async function toKernelSmartAccount<
                 message,
                 accountAddress: await this.getAddress(),
                 kernelVersion: kernelVersion,
-                chainId: await getMemoizedChainId()
+                chainId: await getMemoizedChainId(),
+                eip7702: eip7702
             })
 
             if (isKernelV2(kernelVersion)) {
@@ -741,13 +752,19 @@ export async function toKernelSmartAccount<
             }
 
             return concatHex([
-                getEcdsaRootIdentifierForKernelV3(validatorAddress),
+                getEcdsaRootIdentifierForKernelV3(validatorAddress, eip7702),
                 signature
             ])
         },
         async signTypedData(typedData) {
-            if (eip7702) {
-                throw new Error("Kernel with EIP-7702 isn't 1271 compliant")
+            if (
+                "isDeployed" in this &&
+                !(await (this as any).isDeployed()) &&
+                eip7702
+            ) {
+                throw new Error(
+                    "Kernel with EIP-7702 isn't 1271 compliant before delegation."
+                )
             }
 
             const signature = await signTypedData({
@@ -755,7 +772,8 @@ export async function toKernelSmartAccount<
                 chainId: await getMemoizedChainId(),
                 ...(typedData as TypedDataDefinition),
                 accountAddress: await this.getAddress(),
-                kernelVersion: kernelVersion
+                kernelVersion: kernelVersion,
+                eip7702
             })
 
             if (isKernelV2(kernelVersion)) {
@@ -763,7 +781,7 @@ export async function toKernelSmartAccount<
             }
 
             return concatHex([
-                getEcdsaRootIdentifierForKernelV3(validatorAddress),
+                getEcdsaRootIdentifierForKernelV3(validatorAddress, eip7702),
                 signature
             ])
         },
@@ -789,7 +807,8 @@ export async function toKernelSmartAccount<
                       message: { raw: hash },
                       chainId,
                       accountAddress: await this.getAddress(),
-                      kernelVersion
+                      kernelVersion,
+                      eip7702: false
                   })
                 : await owner.signMessage({
                       message: { raw: hash }